api/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.28.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
httpcore==1.0.9
pydantic_core==2.18.2
openai==1.33.0
